ai/requirements.txt
ultralytics>=8.3.0
supervision>=0.22.0
opencv-python>=4.9.0
httpx>=0.27.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
tqdm>=4.66.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
huggingface_hub>=0.23.0
